requirements.txt
# AgenticTravelRAG Requirements
# Python 3.9+ required

# Core
python-dotenv==1.0.0
pyyaml==6.0.1
typing-extensions==4.8.0

# LangChain & LangGraph
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.2.0

# LLM
openai==1.12.0

# ElasticSearch & Embeddings
elasticsearch==8.11.0
sentence-transformers==2.2.2

# Data Processing
datasets==2.14.0
pandas==2.0.3
numpy==1.24.3
huggingface-hub==0.19.4

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.0

# UI
streamlit==1.28.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging & Monitoring
loguru==0.7.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.0

